requirements.txt
asgiref==3.5.2
autopep8==1.7.0
click==8.1.3
colorama==0.4.5
Django==4.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
h11==0.13.0
pycodestyle==2.9.1
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
uvicorn==0.18.2
